Initial Setup + CV Design
Initial setup of react application is done. It's divided into two parts.
First is Details, which is yet to be developed, but it'll be the section where user will be putting his details that'll be shown in CV.
Second in CV, where all the details user put will be shown in a CV format.
The design of this CV is done.